Refactor file creation to support directories
The code has been updated to allow not just creation of empty files, but also directories. The functions `create_empty_files` and `create_empty_file` have been renamed and refactored to `create_empty_files_or_directories` and `create_empty_file_or_directory` respectively. The new implementation checks if the filename provided has an extension and creates a directory if not.