Add command-line arguments and config file support
Additional command-line arguments were introduced for enhanced configurability. User can now specify a configuration file as an argument. If attributes are not given as command-line arguments, they are fetched from this configuration file. In addition, the help messages are improved to specify that command-line argument settings will override those set in the configuration file.