Refactor logger setup and its usage in functions
The setup_logger function has been removed from individual files and delegated to the newly created class `EnvironmentManager` in the environment.py file. This centralizes the logger setup and prevents the propagation of the logger variable through many function calls. Logger is now passed as a parameter to the functions requiring it. The changes also include refactoring the main script to use `EnvironmentManager`.