Refactor logger initialization and argument setting in EnvironmentManager
This commit refactors the EnvironmentManager class, specifically focusing on the initialization of the logger and the setting of arguments. Initialization messages have been added for both the logger and the EnvironmentManager to improve transparency. The order of operations has been rearranged for argument setting to enhance the logger's performance. Logging levels have also been adjusted accordingly based on the verbosity flag.